+++ crates/apollo-compiler/src/database/ast.rs
@@ -52,17 +52,31 @@ fn syntax_errors(db: &dyn AstDatabase) -> Vec<ApolloDiagnostic> {
             db.ast(file_id)
                 .errors()
                 .map(|err| {
-                    ApolloDiagnostic::new(
-                        db,
-                        (file_id, err.index(), err.data().len()).into(),
-                        DiagnosticData::SyntaxError {
-                            message: err.message().into(),
-                        },
-                    )
-                    .label(Label::new(
-                        (file_id, err.index(), err.data().len()),
-                        err.message(),
-                    ))
+                    if err.is_limit() {
+                        ApolloDiagnostic::new(
+                            db,
+                            (file_id, err.index(), err.data().len()).into(),
+                            DiagnosticData::LimitExceeded {
+                                message: err.message().into(),
+                            },
+                        )
+                        .label(Label::new(
+                            (file_id, err.index(), err.data().len()),
+                            err.message(),
+                        ))
+                    } else {
+                        ApolloDiagnostic::new(
+                            db,
+                            (file_id, err.index(), err.data().len()).into(),
+                            DiagnosticData::SyntaxError {
+                                message: err.message().into(),
+                            },
+                        )
+                        .label(Label::new(
+                            (file_id, err.index(), err.data().len()),
+                            err.message(),
+                        ))
+                    }
                 })
                 .collect::<Vec<ApolloDiagnostic>>()
         })
@@ -72,14 +86,16 @@ fn syntax_errors(db: &dyn AstDatabase) -> Vec<ApolloDiagnostic> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::ApolloCompiler;
+    use crate::{ApolloCompiler, HirDatabase};
 
     #[test]
     fn it_errors_when_selection_set_recursion_limit_exceeded() {
         let schema = r#"
         query {
           Q1 {
-            url
+            url {
+              hostname
+            }
           }
         }
         "#;
@@ -90,7 +106,7 @@ mod tests {
 
         assert_eq!(ast.recursion_limit().high, 2);
         assert_eq!(ast.errors().len(), 1);
-        assert_eq!(ast.document().definitions().count(), 2);
+        assert_eq!(ast.document().definitions().count(), 1);
     }
 
     #[test]
@@ -98,7 +114,11 @@ mod tests {
         let schema = r#"
         query {
           Q1 {
-            url
+            Q2 {
+              Q3 {
+                url
+              }
+            }
           }
         }
         "#;
@@ -170,7 +190,31 @@ mod tests {
         assert_eq!(ast.errors().len(), 1);
         assert_eq!(
             ast.errors().next(),
-            Some(&apollo_parser::Error::limit("parser limit(3) reached", 61))
+            Some(&apollo_parser::Error::limit("parser limit(3) reached", 121))
         );
+    }
+
+    #[test]
+    fn token_limit_with_multiple_sources() {
+        let schema = r#"
+type Query {
+    website: URL,
+    amount: Int
+}
+
+scalar URL @specifiedBy(url: "a.com");
+"#;
+        let query = "{ website }";
+
+        let mut compiler = ApolloCompiler::new();
+        compiler.add_type_system(schema, "schema.graphql");
+        let ts = compiler.db.type_system();
+
+        let mut compiler2 = ApolloCompiler::new().token_limit(2);
+        compiler2.set_type_system_hir(ts);
+        compiler2.add_executable(query, "query.graphql");
+        let parser_errors = compiler2.db.syntax_errors();
+
+        assert_eq!(parser_errors.len(), 1);
     }
 }