+++ crates/apollo-compiler/src/database/hir.rs
@@ -420,6 +420,19 @@ impl FragmentDefinition {
     pub fn loc(&self) -> HirNodeLocation {
         self.loc
     }
+
+    /// Get the location information for the "head" of the fragment definition, namely the
+    /// `fragment` keyword and the name.
+    pub(crate) fn head_loc(&self) -> HirNodeLocation {
+        self.name_src()
+            .loc()
+            .map(|name_loc| HirNodeLocation {
+                // Adjust the node length to include the name
+                node_len: name_loc.end_offset() - self.loc.offset(),
+                ..self.loc
+            })
+            .unwrap_or(self.loc)
+    }
 }
 
 #[derive(Clone, Debug, Hash, PartialEq, Eq)]